+++ src/components/Badge/__mocks__/badge.mock.js
@@ -1,4 +1,4 @@
-import { BADGE_TYPES } from "lib/utils";
+import { BADGE_TYPES } from "lib/utils/badgeTypes";
 
 export default {
   type: BADGE_TYPES.SALE,
