+++ src/components/ProductItem/ProductItem.js
@@ -8,7 +8,7 @@ import Typography from "@material-ui/core/Typography";
 import LoadingIcon from "mdi-material-ui/Loading";
 import Link from "components/Link";
 import BadgeOverlay from "components/BadgeOverlay";
-import { priceByCurrencyCode } from "lib/utils";
+import priceByCurrencyCode from "lib/utils/priceByCurrencyCode";
 import { styles } from "./styles";
 
 @withStyles(styles, { withTheme: true })
