+++ src/components/VariantItem/VariantItem.js
@@ -4,7 +4,7 @@ import { withStyles } from "@material-ui/core/styles";
 import ButtonBase from "@material-ui/core/ButtonBase";
 import Typography from "@material-ui/core/Typography";
 import classNames from "classnames";
-import { priceByCurrencyCode } from "lib/utils";
+import priceByCurrencyCode from "lib/utils/priceByCurrencyCode";
 
 import styles from "./styles";
