+++ src/components/VariantList/VariantList.js
@@ -4,7 +4,7 @@ import { withStyles } from "@material-ui/core/styles";
 import VariantItem from "components/VariantItem";
 import ProductDetailOptionsList from "components/ProductDetailOptionsList";
 import BadgeOverlay from "components/BadgeOverlay";
-import { badgeStatus } from "lib/utils";
+import badgeStatus from "lib/utils/badgeStatus";
 import Divider from "components/Divider";
 
 const styles = (theme) => ({
