+++ src/lib/utils/badgeTypes.js
@@ -1,0 +1,9 @@
+import keyMirror from "keymirror";
+
+export const BADGE_TYPES = keyMirror({
+  BACKORDER: null,
+  BESTSELLER: null,
+  LOW_QUANTITY: null,
+  SOLD_OUT: null,
+  SALE: null
+});
