+++ src/lib/utils/isProductLowQuantity.js
@@ -1,0 +1,9 @@
+/**
+ * Determines if a product has low inventory.
+ *
+ * @param {Object} product - The product
+ * @returns {Boolean} - Indicates whether the product has low inventory
+ */
+export function isProductLowQuantity(product) {
+  return product.isLowQuantity && !product.isSoldOut;
+}
