+++ src/lib/utils/isProducteBestseller.js
@@ -1,0 +1,20 @@
+/**
+ * Determines if a product is a best seller.
+ * TODO: this is a placeholder, as we don't have "Best Seller" at this moment
+ * https://github.com/reactioncommerce/reaction-next-starterkit/issues/130
+ *
+ * @param {Object} product - The product
+ * @returns {Boolean} - Indicates whether the product is a best seller
+ */
+export function isProductBestseller(product) {
+  // This is a placeholder.
+  // Remove the placeholder check and only `return product.isBestseller`
+  // when that functionality if available
+  const placeholder = true;
+  if (placeholder) {
+    return false;
+  }
+
+
+  return product.isBestseller;
+}
