+++ src/lib/utils/pricingByCurrencyCode.js
@@ -1,0 +1,12 @@
+/**
+ * Determines the pricing a product given a currency code.
+ *
+ * @param {String} currencyCode - the shop's set currency code, i.e. "USD"
+ * @param {Array} pricing - An array of pricing objects with different currencies.
+ * @returns {Object} - pricing object, or null if none found.
+ */
+export function priceByCurrencyCode(currencyCode, pricing) {
+  const _pricing = pricing.find((price) => price.currency.code === currencyCode);
+
+  return _pricing || null;
+}
