+++ src/lib/utils/variantById.js
@@ -1,0 +1,14 @@
+/**
+ * Find a variant by _id
+ *
+ * @param {Array} variants - A list of variants
+ * @param {String} variantId - _id of target variant
+ * @returns {Object} - target variant or null if none found.
+ */
+export function variantById(variants, variantId) {
+  const _variant = variants.find((variant) => variant._id === variantId);
+
+  return _variant || null;
+}
+
+
